requirements.txt
beautifulsoup4==4.6.0
pandas==1.1.1
python-docx==0.8.11
selenium==3.141.0